Fix Triton bf16 sigmoid cast, benchmark: 1.31x adapter speedup
Adapter computation: 1.24ms → 0.95ms (31% faster with Triton).
But adapter is only 7% of total decode (16.3ms), so net impact is +1.8%.
Real bottleneck is base model memory bandwidth, not adapter compute.

+++ grove_server/engine/triton_kernels.py
@@ -37,7 +37,8 @@ if HAS_TRITON:
         adapter = tl.load(adapter_ptr + offsets, mask=mask)
         gate_raw = tl.load(gate_ptr + offsets, mask=mask)
 
-        gate = tl.sigmoid(gate_raw)
+        # Triton sigmoid requires fp32
+        gate = tl.sigmoid(gate_raw.to(tl.float32)).to(base.dtype)
         result = base + gate * adapter
         tl.store(out_ptr + offsets, result, mask=mask)
